pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "xnlogo"
version = "0.1.0"
description = "Python-to-NetLogo transpiler and integration toolkit"
authors = [{ name = "Abdulrasheed Ibrahim" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "jinja2>=3.1",
    "mypy>=1.11",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "requests>=2.32",
    "ruff>=0.6",
    "typer>=0.12",
]

[project.optional-dependencies]
runtime = ["jpype1>=1.5"]

dev = ["pytest>=8.3", "pytest-cov>=5.0"]

[project.scripts]
xnlogo = "xnlogo.cli.commands:app"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["xnlogo*"]

[tool.setuptools.package-data]
xnlogo = ["codegen/templates/*.j2"]

[tool.typer]
auto_envvar_prefix = "XNLOGO"
